Begin testing; create test file for npc_combat and tests for get_mod_message()
