+++ main.py
@@ -1,4 +1,4 @@
-from npc_combat import get_attack_value, get_blk_value, get_mod_value
+from npc_combat import get_attack_value, get_blk_value, get_mod_value, get_mod_message
 from combat_card import get_combat_cards
 import random
 
@@ -10,14 +10,6 @@ import random
 ●  	CALCULATION STEP - Determine what happened in this clash. Subtract applicable HP.
 ●  	ACTIVATION PHASE - This is when you activate technique cards that will be played in the next clash. No more than 1 Technique per player may be played in a single clash.
 '''
-
-
-def get_mod_message(mod_value):
-    if mod_value == "Wolf":
-        return "Defense is reversal."
-    if mod_value == "Star":
-        return "Defense is reversal."
-    return "No modification."
 
 
 def clash(pa, pb, ph, ka, kb, kh):
